reconfigured RedundantInfo class in calib.py
We now do not divde/multiply gains out in pack_calpar/unpack_calpar/order_data to go
back to the way things used to be. This is because omnical (in C) already does the division
of the the gains from the data.

Copied back in the old code for order_data to calib.py.